move parity trace types into single crate
