+++ bins/conflux/Cargo.toml
@@ -15,7 +15,7 @@ toml = "0.5.8"
 serde = "1.0"
 serde_json = "1.0"
 serde_derive = "1.0"
-parking_lot = "0.11"
+parking_lot = { workspace = true }
 panic_hook = { git = "https://github.com/Conflux-Chain/conflux-parity-deps.git", rev = "1597a9cab02343eb2322ca0ac58d39b64e3f42d1"  }
 app_dirs = "1.2.1"
 dir = { git = "https://github.com/Conflux-Chain/conflux-parity-deps.git", rev = "1597a9cab02343eb2322ca0ac58d39b64e3f42d1"  }
