+++ crates/accounts/Cargo.toml
@@ -12,7 +12,7 @@ cfxkey = { workspace = true }
 cfxstore = { workspace = true }
 cfx-types = { workspace = true }
 log = "0.4"
-parking_lot = "0.11"
+parking_lot = { workspace = true }
 serde = "1.0"
 serde_derive = "1.0"
 serde_json = "1.0"
