+++ crates/cfx_utils/Cargo.toml
@@ -8,4 +8,4 @@ edition = "2021"
 
 [dependencies]
 log = "0.4"
-parking_lot = "0.11"
+parking_lot = { workspace = true }
