+++ crates/cfxcore/core/Cargo.toml
@@ -58,7 +58,7 @@ network = { workspace = true }
 num = "0.2"
 num-traits = { version = "0.2.8", default-features = false }
 parity-crypto = "0.9.0"
-parking_lot = "0.11"
+parking_lot = { workspace = true }
 primal = "0.2.3"
 primitives = { workspace = true }
 priority-send-queue = { workspace = true }
