+++ crates/cfxcore/core/benchmark/consensus/Cargo.toml
@@ -12,7 +12,7 @@ cfx-types = { workspace = true }
 primitives = { workspace = true }
 db = { workspace = true }
 threadpool = "1.0"
-parking_lot = "0.11"
+parking_lot = { workspace = true }
 log4rs = { version = "1.2.0", features = ["background_rotation", "gzip"] }
 log = "0.4"
