+++ crates/cfxcore/core/src/pos/crypto/crypto/Cargo.toml
@@ -39,7 +39,7 @@ cfx-types = { workspace = true }
 bls-signatures = { workspace = true }
 vrf = "0.2.2"
 lazy_static = "1.4"
-parking_lot = "0.11"
+parking_lot = { workspace = true }
 openssl = "0.10"
 diem-logger = {path = "../../common/logger" }
 pkcs8 = {version = "0.7.5", features = ["encryption", "std"]}
