+++ crates/cfxcore/execute-helper/src/lib.rs
@@ -1,11 +1,8 @@
 #[macro_use]
 extern crate log;
-#[macro_use]
-extern crate error_chain;
 
 pub mod estimation;
 pub mod observer;
 pub mod phantom_tx;
 pub mod tx_outcome;
-
 pub use observer::exec_tracer;
