+++ crates/cfxcore/executor/Cargo.toml
@@ -21,25 +21,25 @@ cfx-vm-types = { workspace = true }
 cfx-vm-interpreter = { workspace = true }
 error-chain = { version = "0.12", default-features = false }
 keccak-hash = { workspace = true }
-lazy_static = "1.4"
-log = "0.4"
+lazy_static = { workspace = true }
+log = { workspace = true }
 malloc_size_of = { workspace = true }
 malloc_size_of_derive = { workspace = true }
 num = "0.2"
 parity-crypto = "0.9.0"
-parking_lot = "0.11"
+parking_lot = { workspace = true }
 primitives = { workspace = true }
 rlp ={ workspace = true }
 rlp_derive = { workspace = true }
-rustc-hex = "2.1"
-serde = { version = "1.0", features = ["rc"] }
-serde_derive = "1.0"
-serde_json = "1.0"
+rustc-hex = { workspace = true }
+serde = { workspace = true, features = ["rc"] }
+serde_derive = { workspace = true }
+serde_json = { workspace = true }
 solidity-abi = { workspace = true }
 solidity-abi-derive = { workspace = true }
 sha3-macro = { workspace = true }
-strum = "0.20"
-strum_macros = "0.20"
+strum = { workspace = true }
+strum_macros = { workspace = true }
 bls-signatures = {git = "https://github.com/Conflux-Chain/bls-signatures.git", rev = "fb52187df92d27c365642cb7e7b2aaf60437cf9c", default-features = false, features = ["multicore"]}
 tiny-keccak = {version = "2.0",  features = ["keccak"]}
 diem-crypto = { path = "../core/src/pos/crypto/crypto" }
@@ -52,7 +52,8 @@ hex-literal = "0.4.1"
 derive_more = "0.99"
 c-kzg = { version = "1.0.2", default-features = false}
 once_cell = "1.19"
-rayon = "1.10"
+rayon = { workspace = true }
+cfx-parity-trace-types = { workspace = true }
 
 [dev-dependencies]
 cfx-statedb = { workspace = true, features = ["testonly_code"]}
