+++ crates/cfxcore/parity-trace-types/src/lib.rs
@@ -1,0 +1,10 @@
+pub mod action_types;
+pub mod filter;
+pub mod trace_types;
+
+#[cfg(test)]
+mod tests;
+
+pub use action_types::*;
+pub use filter::*;
+pub use trace_types::*;
