+++ crates/cfxcore/vm-types/Cargo.toml
@@ -14,8 +14,8 @@ keccak-hash = { workspace = true }
 primitives = { workspace = true }
 rlp = { workspace = true }
 rlp_derive = { workspace = true }
-serde = { version = "1.0", features = ["rc"] }
-serde_derive = "1.0"
+serde = { workspace = true, features = ["rc"] }
+serde_derive = { workspace = true }
 solidity-abi = { workspace = true }
 
 [features]
