+++ crates/dbs/kvdb-rocksdb/Cargo.toml
@@ -13,7 +13,7 @@ fs-swap = "0.2.4"
 kvdb = "0.4"
 log = "0.4.8"
 num_cpus = "1.10.1"
-parking_lot = "0.11"
+parking_lot = { workspace = true }
 regex = "1.3.1"
 malloc_size_of = { workspace = true }
 malloc_size_of_derive = { workspace = true }
