+++ crates/dbs/storage/Cargo.toml
@@ -28,7 +28,7 @@ log4rs = { version = "1.2.0", features = ["background_rotation", "gzip"] }
 malloc_size_of = { workspace = true }
 malloc_size_of_derive = { workspace = true }
 memoffset = "0.5.1"
-parking_lot = "0.11"
+parking_lot = { workspace = true }
 primitives = { workspace = true, optional = true }
 rand = "0.7"
 rand_chacha="0.2.1"
