+++ crates/network/Cargo.toml
@@ -9,7 +9,7 @@ cfx-bytes = { workspace = true }
 cfx-types = { workspace = true }
 cfxkey = { workspace = true }
 mio = "0.6.8"
-parking_lot = "0.11"
+parking_lot = { workspace = true }
 log = "0.4"
 slab = "0.4"
 io = { workspace = true }
