+++ crates/secret_store/Cargo.toml
@@ -8,7 +8,7 @@ edition = "2021"
 
 [dependencies]
 cfx-types = { workspace = true }
-parking_lot = "0.11"
+parking_lot = { workspace = true }
 rlp = { workspace = true }
 rand = "0.7"
 cfxkey = { workspace = true }
