+++ crates/stratum/Cargo.toml
@@ -12,7 +12,7 @@ keccak-hash = { workspace = true }
 jsonrpc-core = "15.1.0"
 jsonrpc-tcp-server = "15.1.0"
 log = "0.4"
-parking_lot = "0.11"
+parking_lot = { workspace = true }
 
 [dev-dependencies]
 env_logger = "0.5"
