+++ crates/transactiongen/Cargo.toml
@@ -13,7 +13,7 @@ cfx-types = { workspace = true }
 cfx-vm-types = { workspace = true }
 cfx-bytes = { workspace = true }
 primitives = { workspace = true }
-parking_lot = "0.11"
+parking_lot = { workspace = true }
 rlp = { workspace = true }
 rand = "0.7"
 rustc-hex = "2.1"
