+++ crates/util/io/Cargo.toml
@@ -11,7 +11,7 @@ fnv = "1.0"
 mio = { version = "0.6.19" }
 crossbeam-deque = "0.7"
 crossbeam-channel = "0.4"
-parking_lot = "0.11"
+parking_lot = { workspace = true }
 log = "0.4"
 slab = "0.4"
 num_cpus = "1.8"
