+++ crates/util/link-cut-tree/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2018"
 
 [dependencies]
 cfx-types = { workspace = true }
-parking_lot = "0.11"
+parking_lot = { workspace = true }
 rand = "0.7"
 malloc_size_of = { workspace = true }
 malloc_size_of_derive = { workspace = true }
