+++ crates/util/malloc_size_of/Cargo.toml
@@ -10,7 +10,7 @@ cfg-if = "0.1"
 hashbrown = "0.7.1"
 winapi = "0.3.7"
 slab = "0.4"
-parking_lot = "0.11"
+parking_lot = { workspace = true }
 smallvec = "1.4"
 
 [features]
