+++ crates/util/metrics/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2018"
 
 [dependencies]
 lazy_static = "1.4"
-parking_lot = "0.11"
+parking_lot = { workspace = true }
 timer = "0.2.0"
 time = "0.1"
 rand = "0.7"
