+++ crates/util/throttling/Cargo.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 edition = "2018"
 
 [dependencies]
-parking_lot = "0.11"
+parking_lot = { workspace = true }
 toml = "0.5.8"
 malloc_size_of = { workspace = true }
 malloc_size_of_derive = { workspace = true }
